chore(bonds): organise files and folders
