--- src/components/ComponentTemplates/Stateful.tsx
@@ -1,13 +1,0 @@
-import * as React from 'react'
-
-export interface ParentProps {
-  title: string
-}
-
-export class Stateful extends React.Component<ParentProps, {}> {
-  state = {}
-
-  render(): JSX.Element {
-    return <div className="container-fluid">CONTENT HERE</div>
-  }
-}
