--- src/components/ComponentTemplates/Stateless.tsx
@@ -1,9 +1,0 @@
-import * as React from 'react'
-
-export interface ParentProps {
-  title: string
-}
-
-export const Stateless: React.SFC<ParentProps> = () => {
-  return <div className="container-fluid">test</div>
-}
