--- src/components/Token.scss
@@ -1,20 +1,0 @@
-@import './../assets/styles/colors';
-@import './../assets/styles/fonts';
-
-.Token {
-  border-radius: 0.375rem;
-  padding: 0.25rem 0.688rem 0.25rem 0.688rem;
-  font-size: $token-font-size;
-  font-weight: bold;
-  margin-right: 0.5rem;
-}
-
-.Token_EDU {
-  background-color: $token-edu-background-color;
-  color: $token-edu-color;
-}
-
-.Token_IXO {
-  background-color: $token-ixo-background-color;
-  color: $token-ixo-color;
-}
