+++ src/components/bonds/overview/charts/Charts.tsx
@@ -1,0 +1,38 @@
+import React, { Component } from 'react'
+import { Bar } from 'react-chartjs-2'
+
+const data = {
+  labels: ['January', 'February', 'March', 'April', 'May', 'June', 'July'],
+  datasets: [
+    {
+      label: 'My First dataset',
+      backgroundColor: 'rgba(255,99,132,0.2)',
+      borderColor: 'rgba(255,99,132,1)',
+      borderWidth: 1,
+      hoverBackgroundColor: 'rgba(255,99,132,0.4)',
+      hoverBorderColor: 'rgba(255,99,132,1)',
+      data: [65, 59, 80, 81, 56, 55, 40],
+    },
+  ],
+}
+
+export class Charts extends Component {
+  render() {
+    return (
+      <div className="BondsWrapper_panel__chrome">
+        <div className="BondsWrapper_panel__content">
+          <b>Charts</b>
+
+          <Bar
+            data={data}
+            width={100}
+            height={500}
+            options={{
+              maintainAspectRatio: false,
+            }}
+          />
+        </div>
+      </div>
+    )
+  }
+}
