+++ src/components/bonds/overview/events/Events.tsx
@@ -1,0 +1,13 @@
+import React, { Component } from 'react'
+
+export class Events extends Component {
+  render() {
+    return (
+      <div className="BondsWrapper_panel__chrome">
+        <div className="BondsWrapper_panel__content">
+          <b>Events</b>
+        </div>
+      </div>
+    )
+  }
+}
