--- src/components/header/AppHead.scss
@@ -1,24 +1,0 @@
-@import './../../assets/styles/colors';
-
-.AppHead_header {
-  background: $header-background-color;
-  background: $header-background-color-gradient;
-  padding: 1rem;
-  margin: 2rem 2rem 2rem -2px;
-  z-index:1;
-  border: 1px solid $header-border-color;
-  display: flex;
-  flex-wrap: wrap;
-  flex-direction: row;
-  justify-content: left;
-/*    font-size: calc(0.625rem + 2vmin);
-    color: white;*/
-}
-
-  /*.AppHead_header__text {
-    color: lightsteelblue;
-  }
-
-  .AppHead_header__textUpper {
-    text-transform: uppercase;
-  }*/
