--- src/components/overview/events/Events.tsx
@@ -1,14 +1,0 @@
-import React, { Component } from 'react';
-
-export class Events extends Component {
-   render() {
-      return (
-         <div className="AppBody_panel__chrome" >
-
-            <div className="AppBody_panel__content">
-               <b>Events</b>
-            </div>
-         </div>
-      );
-   }
-}
